requirements.txt
aioredis==1.3.1
apturl==0.5.2
asgiref==3.3.1
astroid==2.5.2
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.6
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
chardet2==2.0.3
charset-normalizer==2.0.1
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.7
cupshelpers==1.0
daphne==3.0.2
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.7
djangorestframework==3.12.4
entrypoints==0.3
flake8==3.9.0
future==0.18.2
gocardless-pro==1.22.0
hiredis==2.0.0
httplib2==0.14.0
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
isort==5.8.0
jedi==0.17.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
louis==3.12.0
macaroonbakery==1.3.1
mccabe==0.6.1
msgpack==1.0.2
netifaces==0.10.4
oauth2==1.9.0.post1
oauthlib==3.1.0
olefile==0.46
parso==0.7.1
pexpect==4.6.0
Pillow==7.0.0
pluggy==0.13.1
protobuf==3.6.1
psutil==5.5.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycups==1.9.73
pydocstyle==6.0.0
pyflakes==2.2.0
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.7.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2019.3
PyVimeo==1.1.0
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.26.0
requests-unixsocket==0.2.0
rope==0.18.0
SecretStorage==2.3.1
service-identity==21.1.0
simplejson==3.16.0
six==1.14.0
snowballstemmer==2.1.0
sqlparse==0.4.1
stripe==2.59.0
systemd-python==234
tinydb==4.5.1
toml==0.10.2
tuspy==0.2.5
Twisted==21.2.0
txaio==21.2.1
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==4.0.2
unattended-upgrades==0.1
urllib3==1.25.8
vimeo==0.1.9
wadllib==1.3.3
wrapt==1.12.1
xkit==0.0.0
yapf==0.31.0
zope.interface==5.4.0
